+++ tests/__init__.py
@@ -3,3 +3,5 @@ from .test_parse import *
 from .test_encoding_error import *
 from .test_encode import *
 from .test_decode import *
+from .test_hash_field import *
+from .test_hash_raw import *